hr-backend/pyproject.toml
[project]
name = "hr-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==25.1.0",
    "aiomysql==0.3.2",
    "aiosmtplib==5.0.0",
    "alembic==1.17.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "apscheduler==3.11.2",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.4.0",
    "asgiref==3.11.0",
    "asttokens==3.0.1",
    "async-lru==2.1.0",
    "attrs==25.4.0",
    "babel==2.17.0",
    "beautifulsoup4==4.14.3",
    "bleach==6.3.0",
    "blinker==1.9.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==46.0.3",
    "cycler==0.12.1",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "et-xmlfile==2.0.0",
    "executing==2.2.1",
    "fastapi==0.124.2",
    "fastapi-cache2[redis]==0.2.2",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.6.0",
    "fastapi-mail==1.6.0",
    "fastar==0.8.0",
    "fastjsonschema==2.21.2",
    "ffmpeg-python==0.2.0",
    "filelock==3.20.0",
    "fonttools==4.61.1",
    "fqdn==1.5.1",
    "future==1.0.0",
    "grandalf==0.8",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "ipykernel==7.1.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "json5==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.3.0",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.4",
    "jupyterlab==4.5.3",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.28.0",
    "kiwisolver==1.4.9",
    "langchain==1.2.0",
    "langchain-core==1.2.2",
    "langchain-deepseek==1.0.1",
    "langchain-openai==1.1.4",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-checkpoint-mysql==2.0.17",
    "langgraph-checkpoint-postgres==3.0.2",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.0",
    "langsmith==0.5.0",
    "lark==1.3.1",
    "loguru==0.7.3",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mistune==3.2.0",
    "modelscope==1.33.0",
    "nbclient==0.10.4",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.5.3",
    "notebook-shim==0.2.4",
    "numpy==2.4.1",
    "openai==2.13.0",
    "openpyxl==3.1.5",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "pandas==3.0.0",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pendulum==3.1.0",
    "pillow==12.1.0",
    "platformdirs==4.5.1",
    "pluggy==1.6.0",
    "prometheus-client==0.24.1",
    "prompt-toolkit==3.0.52",
    "psutil==7.2.1",
    "psycopg==3.3.2",
    "psycopg-binary==3.3.2",
    "psycopg-pool==3.3.0",
    "pure-eval==0.2.3",
    "pwdlib==0.3.0",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pymupdf==1.26.7",
    "pymysql==1.1.2",
    "pyparsing==3.3.2",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.20",
    "pywin32>=311",
    "pywinpty==3.0.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rich==14.2.0",
    "rich-toolkit==0.17.0",
    "rignore==0.7.6",
    "rpds-py==0.30.0",
    "send2trash==2.1.0",
    "sentry-sdk==2.47.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "shortuuid==1.0.13",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "starlette==0.50.0",
    "tenacity==9.1.2",
    "terminado==0.18.1",
    "tiktoken==0.12.0",
    "tinycss2==1.4.0",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "uri-template==1.3.0",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "uv==0.9.17",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "wcwidth==0.5.0",
    "webcolors==25.10.0",
    "webencodings==0.5.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "win32-setctime==1.2.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
